all build work done by Makefile not with the help of SHELL script
