--- boot_nand/README.txt
@@ -1,4 +1,0 @@
-thoes files in this folder is used for booting from nand flash.
-
-you could use dnw and u-boot(enter 'n' in menu) to load the image into nand flash.
-
