--- boot_nand/boot.h
@@ -1,1 +1,0 @@
-/* nothing here */
