--- boot_nor/README.txt
@@ -1,6 +1,0 @@
-thoes files in this folder is used for booting from nor flash while the default
-booting method is nand flash.
-
-you could use dnw(remember to set the download address 0x30080000) and
-u-boot(enter 'd' in menu) to load the image into sdram.
-
