+++ tools/dnw/win7-64-DNW-USB/secbulk64/remove.bat
