+++ tools/prebuild/bare2440_boot_from_nor.bin
